+++ src/main/java/c4/conarm/common/armor/modifiers/ModConcealed.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2018 <C4>
+ *
+ * This Java class is distributed as a part of the Construct's Armory mod.
+ * Construct's Armory is open source and distributed under the GNU Lesser General Public License v3.
+ * View the source code and license file on github: https://github.com/TheIllusiveC4/ConstructsArmory
+ *
+ * Some classes and assets are taken and modified from the parent mod, Tinkers' Construct.
+ * Tinkers' Construct is open source and distributed under the MIT License.
+ * View the source code on github: https://github.com/SlimeKnights/TinkersConstruct/
+ * View the MIT License here: https://tldrlegal.com/license/mit-license
+ */
+
+package c4.conarm.common.armor.modifiers;
+
+import c4.conarm.lib.modifiers.ArmorModifier;
+import net.minecraft.nbt.NBTTagCompound;
+import slimeknights.tconstruct.library.modifiers.ModifierAspect;
+
+public class ModConcealed extends ArmorModifier {
+
+    public ModConcealed() {
+        super("concealed", 0xfff5cc);
+
+        addAspects(new ModifierAspect.SingleAspect(this), new ModifierAspect.DataAspect(this));
+    }
+
+    @Override
+    public void applyEffect(NBTTagCompound rootCompound, NBTTagCompound modifierTag) {
+        //NO-OP
+    }
+}